pyproject.toml
[tool.poetry]
name = "pymapper"
version = "0.1.0"
description = "Python standalone map rendering library using standard python libraries and pycairo"
authors = ["Dolf Andringa <dolfandringa@gmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.8"
geopandas = "^0.10.2"
pyproj = "^3.3.0"
pycairo = "^1.20.1"
OWSLib = "^0.25.0"
Shapely = "^1.8.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pylint = "^2.12.2"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
codecov = "^2.1.12"
flake8 = "^4.0.1"
Sphinx = "^4.4.0"
myst-parser = "^0.17.0"
sphinx-rtd-theme = "^1.0.0"
pytest-pylint = "^0.18.0"
pytest-flake8 = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
preview = true

[tool.pytest.ini_options]
addopts = "--cov=pymapper --cov-report=xml --flake8 --pylint --pylint-rcfile=setup.cfg"

[tool.pyright]
reportUnusedImport = false
reportUnusedClass = false
